+++ docs/README.md
@@ -1,4 +1,4 @@
-# Reboot User Guide
+# reboot.Reboot User Guide
 
 // Update the title above to match the actual product name
