+++ src/main/java/reboot/Parser.java
@@ -1,3 +1,10 @@
+package reboot;
+
+import reboot.command.*;
+import reboot.task.Deadline;
+import reboot.task.Event;
+import reboot.task.Todo;
+
 public class Parser {
 
     public static Command parse(String input) throws RebootException {
