+++ src/main/java/reboot/RebootException.java
@@ -1,3 +1,5 @@
+package reboot;
+
 public class RebootException extends RuntimeException {
     public RebootException(String message) {
         super("Reboot error: " + message);
