+++ src/main/java/reboot/Storage.java
@@ -1,3 +1,10 @@
+package reboot;
+
+import reboot.task.Deadline;
+import reboot.task.Event;
+import reboot.task.Task;
+import reboot.task.Todo;
+
 import java.io.BufferedWriter;
 import java.io.IOException;
 import java.nio.file.Files;
