+++ src/main/java/reboot/command/AddCommand.java
@@ -1,3 +1,12 @@
+package reboot.command;
+
+import reboot.*;
+import reboot.task.Deadline;
+import reboot.task.Event;
+import reboot.task.Task;
+import reboot.task.Todo;
+
+
 public class AddCommand extends Command{
 
     private final Task task;
@@ -21,4 +30,5 @@ public class AddCommand extends Command{
         ui.showMessage("    Updated\n      " + task +
                 "\n    " + tasks.size() + " tasks in the list");
     }
+
 }
