+++ src/main/java/reboot/command/ClearCommand.java
@@ -1,4 +1,9 @@
-public class ClearCommand extends Command{
+package reboot.command;
+
+import reboot.RebootException;
+import reboot.*;
+
+public class ClearCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws RebootException {
