+++ src/main/java/reboot/command/Commands.java
@@ -1,3 +1,5 @@
+package reboot.command;
+
 public enum Commands {
     LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, BYE, CLEAR
 }
