+++ src/main/java/reboot/command/DeleteCommand.java
@@ -1,20 +1,28 @@
-public class UnmarkCommand extends Command{
+package reboot.command;
+
+import reboot.*;
+import reboot.task.Task;
+
+public class DeleteCommand extends Command {
+
     private final int index;
 
-    public UnmarkCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws RebootException {
 
+        // Throw exception when number provided is not within index
         if (tasks.isOutOfRange(index)) {
             throw new RebootException("Only numbers from 1 to " + tasks.size() + " are allowed");
         }
 
         Task t = tasks.get(index - 1);
-        t.unmark();
+        tasks.remove(t);
         storage.writeFile(tasks.getAll());
-        ui.showMessage("    Unmarked\n" + "      " + t);
+        ui.showMessage("    Deleting task\n      " + t +
+                "\n    " + tasks.size() + " tasks in the list\n");
     }
 }
