+++ src/main/java/reboot/command/ExitCommand.java
@@ -1,4 +1,8 @@
-public class ExitCommand extends Command{
+package reboot.command;
+
+import reboot.*;
+
+public class ExitCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws RebootException {
