+++ src/main/java/reboot/task/Todo.java
@@ -1,3 +1,5 @@
+package reboot.task;
+
 public class Todo extends Task {
     public Todo(String description, boolean isDone) {
         super(description, isDone);
